audiofilters, audiodelays: fix silence fill, clipping bound and loop counters
memset(word_buffer, 32768, ...) writes only the low byte of its value, and
32768 & 0xff is 0. So the 'fill with silence' path filled the buffer with
zeros -- which for unsigned samples is full negative deflection, not
silence. It writes 0x8000 words in a loop now.

Distortion clamped with MIN(MAX(word, -32767), 32768). The upper bound
becomes -32768 on the cast to int16_t below it, so a positive overdrive
came out as a full negative spike.

filter_states_len is a size_t but all three loops over it counted with a
uint8_t, so with 256 or more Biquad objects the counter wrapped back to
zero and the loop never terminated.

Not compile-checked: neither module is built in this configuration, and
enabling them temporarily to get a compile pass broke qstr generation. The
buffer types were confirmed by reading (word_buffer is int16_t *, length is
a sample count) but no compiler has seen these edits.

+++ shared-module/audiofilters/Filter.c
@@ -132,7 +132,10 @@ void audiofilters_filter_reset_buffer(audiofilters_filter_obj_t *self,
     memset(self->filter_buffer, 0, SYNTHIO_MAX_DUR * sizeof(int32_t));
 
     if (self->filter_states) {
-        for (uint8_t i = 0; i < self->filter_states_len; i++) {
+        // CIRCUITPY-CHANGE: filter_states_len is a size_t, but all three loops over
+        // it used uint8_t counters. With 256 or more Biquad objects the counter
+        // wrapped back to zero and the loop never terminated.
+        for (size_t i = 0; i < self->filter_states_len; i++) {
             synthio_biquad_filter_reset(&self->filter_states[i]);
         }
     }
@@ -207,7 +210,7 @@ audioio_get_buffer_result_t audiofilters_filter_get_buffer(audiofilters_filter_o
             (void)synthio_block_slot_get(&self->mix);
 
             // Tick biquad filters
-            for (uint8_t j = 0; j < self->filter_states_len; j++) {
+            for (size_t j = 0; j < self->filter_states_len; j++) {
                 common_hal_synthio_biquad_tick(self->filter_objs[j]);
             }
             if (self->base.samples_signed) {
@@ -265,7 +268,7 @@ audioio_get_buffer_result_t audiofilters_filter_get_buffer(audiofilters_filter_o
                     }
 
                     // Process biquad filters
-                    for (uint8_t j = 0; j < self->filter_states_len; j++) {
+                    for (size_t j = 0; j < self->filter_states_len; j++) {
                         mp_obj_t filter_obj = self->filter_objs[j];
                         common_hal_synthio_biquad_tick(filter_obj);
                         synthio_biquad_filter_samples(filter_obj, &self->filter_states[j], self->filter_buffer, n_samples);
